test: move HMR Full Reload test to a separate file
