+++ test/development/basic/hmr.test.ts
@@ -1147,90 +1147,6 @@ describe.each([
     })
   })
 
-  describe('Full reload', () => {
-    it('should warn about full reload in cli output - anonymous page function', async () => {
-      const start = next.cliOutput.length
-      const browser = await webdriver(
-        next.url,
-        basePath + '/hmr/anonymous-page-function'
-      )
-      const cliWarning =
-        'Fast Refresh had to perform a full reload when ./pages/hmr/anonymous-page-function.js changed. Read more: https://nextjs.org/docs/messages/fast-refresh-reload'
-
-      expect(await browser.elementByCss('p').text()).toBe('hello world')
-      expect(next.cliOutput.slice(start)).not.toContain(cliWarning)
-
-      const currentFileContent = await next.readFile(
-        './pages/hmr/anonymous-page-function.js'
-      )
-      const newFileContent = currentFileContent.replace(
-        '<p>hello world</p>',
-        '<p id="updated">hello world!!!</p>'
-      )
-      await next.patchFile(
-        './pages/hmr/anonymous-page-function.js',
-        newFileContent
-      )
-
-      expect(await browser.waitForElementByCss('#updated').text()).toBe(
-        'hello world!!!'
-      )
-
-      // CLI warning
-      expect(next.cliOutput.slice(start)).toContain(cliWarning)
-
-      // Browser warning
-      const browserLogs = await browser.log()
-      expect(
-        browserLogs.some(({ message }) =>
-          message.includes(
-            "Fast Refresh will perform a full reload when you edit a file that's imported by modules outside of the React rendering tree."
-          )
-        )
-      ).toBeTruthy()
-    })
-
-    it('should warn about full reload in cli output - runtime-error', async () => {
-      const start = next.cliOutput.length
-      const browser = await webdriver(next.url, basePath + '/hmr/runtime-error')
-      const cliWarning =
-        'Fast Refresh had to perform a full reload due to a runtime error.'
-
-      await retry(async () => {
-        expect(await getRedboxHeader(browser)).toMatch(
-          /ReferenceError: whoops is not defined/
-        )
-      })
-      expect(next.cliOutput.slice(start)).not.toContain(cliWarning)
-
-      const currentFileContent = await next.readFile(
-        './pages/hmr/runtime-error.js'
-      )
-      const newFileContent = currentFileContent.replace(
-        'whoops',
-        '<p id="updated">whoops</p>'
-      )
-      await next.patchFile('./pages/hmr/runtime-error.js', newFileContent)
-
-      expect(await browser.waitForElementByCss('#updated').text()).toBe(
-        'whoops'
-      )
-
-      // CLI warning
-      expect(next.cliOutput.slice(start)).toContain(cliWarning)
-
-      // Browser warning
-      const browserLogs = await browser.log()
-      expect(
-        browserLogs.some(({ message }) =>
-          message.includes(
-            '[Fast Refresh] performing full reload because your application had an unrecoverable error'
-          )
-        )
-      ).toBeTruthy()
-    })
-  })
-
   if (!process.env.TURBOPACK) {
     it('should have client HMR events in trace file', async () => {
       const traceData = await next.readFile('.next/trace')
